requirements.txt
asgiref==3.3.1
astroid==2.4.2
dj-database-url==0.5.0
Django==3.1.4
django-autoslug==1.9.8
django-awesome-avatar==1.1.3
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-wysiwyg==0.8.0
gunicorn==20.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
